Changed Button that accesses estimote page
Changed the button that triggers the new estimote page and fixed the
error that was causing the ranging text to freeze up

+++ DrU/EstimoteViewController.cs
@@ -120,11 +120,19 @@ namespace DrU
                 Debug.Write("App Support already exists");
             }
 
-            //Setting up new file directory
-            var file = Path.Combine(supDir.Path, "Test.txt");
 
-            //Writing to created file
-            File.WriteAllText(file, "This is a test");
+            if (!NSFileManager.DefaultManager.FileExists(Path.Combine(supDir.Path, "Test.txt")))
+            {
+                //Setting up new file directory
+                var file = Path.Combine(supDir.Path, "Test.txt");
+
+                //Writing to created file
+                File.WriteAllText(file, "This is a test");
+            }
+            else
+            {
+                Debug.Write("File Exists");
+            }
 
             //Checking if file was created
             var checker = NSFileManager.DefaultManager.FileExists(Path.Combine(supDir.Path, "Test.txt"));
